requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.7
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
Werkzeug==0.16.1
